requirements.txt
streamlit
torch>=2.1.2
torchvision>=0.16.2
pillow>=10.3.0
transformers>=4.41.0
peft>=0.11.1
accelerate>=0.29.3
safetensors>=0.4.2
sentencepiece>=0.2.0
